Rename evaluation components to eval for brevity
